Fix tracking rectangles in 64-bit build by remembering the tag as the 64-bit integer that it is in that mode.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_8_BRANCH@51595 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/cocoa/trackingrectmanager.h
@@ -32,7 +32,7 @@ public:
 protected:
     wxWindow *m_window;
     bool m_isTrackingRectActive;
-    int m_trackingRectTag;
+    NSInteger m_trackingRectTag;
     NSRect m_trackingRectInWindowCoordinates;
 private:
     wxCocoaTrackingRectManager();
